Markdonw block: fixes issue pasting Markdown
There was a bug when pasting Markdown code containing any type of list
in the live preview. The Markdown parser, Markdown-It, removes empty
space at the beginning of new lines, making impossible to paste lists
and nested lists. Also, changes the whitespace CSS property for
paragraphs inside the Live Preview component, to `pre-wrap` so white
space and new lines are not collapsed.